fix(web): restructure seat selection component files
