+++ apps/web/src/features/seat-selection/components/seat-selection/seat-map.tsx
@@ -2,8 +2,8 @@ import AutoSizer from "react-virtualized-auto-sizer"
 import { FixedSizeList } from "react-window"
 import { Button } from "@/shared/components/button"
 import { Card } from "@/shared/components/card"
-import { useSeatMetrics } from "../hooks/use-seat-metrics"
-import { useSeatMapStore } from "../store/use-seat-map-store"
+import { useSeatMetrics } from "../../hooks/use-seat-metrics"
+import { useSeatMapStore } from "../../store/use-seat-map-store"
 import { Seat } from "./seat"
 import { SeatRow } from "./seat-row"
